==> 02_scrapping.ipynb <==
# Importation

from bs4 import BeautifulSoup as bs
from urllib.request import urlopen
import requests 
import pandas as pd
import numpy as np

# Scrapping jeuxvideos.com

## Fonction get_data(link)

def get_data(link):

  """
  Fonction qui prend en argument un url du site jeuxvideos.com, rubrique 'tous les jeux'
  et qui renvoit un DataFrame contenant le titre du jeu, les notes de tests et les notes d'avis.
  """
  """
  La fonction différencie la première page des autres car les éléments de scrapping diffèrent sur la première page.
  """

  page = requests.get(link)
  page.text
  soup = bs(page.text, "html.parser")

  titles = []
  test = []
  avis = []

  # Si le lien est celui de la premiere page du site
  if link == 'https://www.jeuxvideo.com/tous-les-jeux/':

    # Titres
    for element in soup.select("h2 > a"):
      titles.append(element.text)

    # Notes test
    for element in soup.find_all('div', class_ = 'gameHub__2Gf7Q4'):
      if element.find_next('span', class_ = 'editorialRating__1tYu_r'):
        test.append(element.find_next('span', class_ = 'editorialRating__1tYu_r').text.strip())
      else :
        test.append('Non renseigné')

    # Notes avis
    for element in soup.find_all('div', class_ = 'gameHub__2Gf7Q4'):
      if element.find_next('span', class_ = 'userRating__1y96su'):
        avis.append(element.find_next('span', class_ = 'userRating__1y96su').text.strip())
      else :
        avis.append('Non renseigné')

    # Création d'un DataFrame
    df = pd.DataFrame({'titres': titles,'test': test, 'avis':avis})
  
  # Sinon le lien correpond à une page ultérieure du type https://www.jeuxvideo.com/tous-les-jeux/?=NUMERO_DE_PAGE

  else :
    
    # Titres
    name = soup.find_all('h2', attrs = {'class': 'gameTitle__3PBZI1'})
    
    for element in name :
      titles.append(element.find('span', attrs = {'class' : 'gameTitleLink__196nPy'}).text)

    # Récupération des notes
    notes = soup.find_all('div', attrs = {'class' : 'gameHub__2Gf7Q4'})

    for element in notes :

      # Notes test
      try :
        if element.find('span', attrs = {'class' : 'editorialRating__1tYu_r'}):
          test.append(element.find('span', attrs = {'class' : 'editorialRating__1tYu_r'}).text)
        elif element.find('span', attrs = {'class' : 'editorialRating__1tYu_r'}) == None : #si le None n'est pas renseigné, l'except n'est pas lu et la liste est raccourcie d'un élément
          test.append('Non renseigné')
      except :
          test.append('Non renseigné')

      # Notes avis
      try :
        if element.find('span', attrs = {'class' : 'userRating__1y96su'}):
          avis.append(element.find('span', attrs = {'class' : 'userRating__1y96su'}).text)
        elif element.find('span', attrs = {'class' : 'userRating__1y96su'}) == None :
          avis.append('Non renseigné')
      except :
          avis.append('Non renseigné')
          
    # Création d'un DataFrame 
    df = pd.DataFrame({'titres' : titles,'test':test,'avis':avis})

  return df

## Fonction get_data_jeu(link)

def get_data_jeu(link):

  """fonction qui prend en argument une page du site jeuxvideos.com référençant un jeu
  et qui renvoit une liste d'information sur ce jeu. Les informations sont :

  - Nom du jeu
  - Note d'avis
  - Note de test
  - Platformes disponibles
  
  Format de l'URL :
  'https://www.jeuxvideo.com/jeux/jeu-1716149/' """

  # Initialisation
  page = requests.get(link)
  page.text
  soup = bs(page.text, "lxml")

  # Instanciation de listes vides
  liste_plateformes = []
  liste_publisher = []
  donnees_jeuxvideos = []

  # Nom du jeu vidéo
  name = soup.find('h1', attrs = {'class' : 'breadcrumb__item'}).text

  # Notes du jeu : instanciation d'un bloc Try pour lever les exceptions (tous les jeus ne sont pas notés)
  try :
    if soup.find_all('text', attrs = {'class' : 'gameCharacteristicsMain__gaugeText'}): # si la partie note existe sur la page
      test = soup.find_all('text', attrs = {'class' : 'gameCharacteristicsMain__gaugeText'})[0].text # on récupère les 2 premiers élements text
      avis = soup.find_all('text', attrs = {'class' : 'gameCharacteristicsMain__gaugeText'})[1].text
  except : # si la partie note n'existe pas, on lève l'exception en renseignant "Non trouvé"
      test = 'Non trouvé'
      avis = 'Non trouvé'

  # Plateforme de jeu
  support = soup.find_all('li', attrs = {'class' : 'gameHeaderBanner__platform'})
  
  for element in support :
    liste_plateformes.append(element.text.strip()) # certaines données contiennent des espaces et des sauts de lignes

  # Publisher
  publisher = soup.find('div', attrs = {'class':"gameCharacteristicsDetailed__tr"})
  span = publisher.find_all('span')

  for element in span :
    liste_publisher.append(element.text.strip())


  donnees_jeuxvideos.append(name)
  donnees_jeuxvideos.append(test)
  donnees_jeuxvideos.append(avis)
  donnees_jeuxvideos.append(liste_plateformes)
  donnees_jeuxvideos.append(liste_publisher)
  
  return donnees_jeuxvideos

## Fonction get_data_n_pages(n)

def get_data_n_pages(n_pages):

  """ Fonction qui prend en argument un nombre de page sur le site jeuxvideos.com
  et qui retourne un DataFrame contenant les informations sur les jeux videos.
  """
  # Instanciation d'un DataFrame vide
  data_jeuxvideos = pd.DataFrame()

  # DataFrame correspondant à la 1ère page
  df1 = get_data('https://www.jeuxvideo.com/tous-les-jeux/')

  # Concaténation du DF de la 1ère page avec le DF vide
  data_jeuxvideos = pd.concat([data_jeuxvideos, df1])

  # Récupération des DF des autres pages et concaténation
  for i in range(2,n_pages,1) :

    link = 'https://www.jeuxvideo.com/tous-les-jeux/?p=' + str(i)
    data_jeuxvideos = pd.concat([data_jeuxvideos, get_data(link)])

  # La fonction renvoit le DataFrame final
  return data_jeuxvideos.reset_index().drop('index', axis = 1)


# Le scrapping du site a eu lieu le 19 avril et le dataframe généré a été enregistré sous format
# CSV sous le nom dataframe_jeuxvideos.csv
# La mise à jour régulière du site jeuxvideos.com peut générer des erreurs

path = "dataframe_jeuxvideos.csv"

data = pd.read_csv(path)
data.head()

data = data.drop('Unnamed: 0', axis = 1)
data.head()

data.info()

# Scrapping metacritic.com

#import des modules de scrapping

from bs4 import BeautifulSoup as bs
from bs4 import UnicodeDammit
import requests
import pandas as pd
import numpy as np


#choisir l'url de metacritic (page "Best video games of all time")
url = 'https://www.metacritic.com/browse/games/score/metascore/all/all/filtered'

#méthode trouvée sur stackoverflow pour résoudre le problème 'requests.exceptions.TooManyRedirects: Exceeded 30 redirects'
page = requests.Session()
page.headers['User-Agent'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/34.0.1847.131 Safari/537.36'
réponse = page.get(url)

#Contenu HTML et réponse: afficher le contenu du site web.
soup = bs(réponse.content,'html.parser')

#Création des listes pour le dataframe

Name = []
Date = []
Platform = []
Note = []
P_Note = []

# Extraction des éléments de la page qui contiennent les données recherchées
game_elements = soup.findAll("td", class_="clamp-summary-wrap")

# Parcourir les éléments pour extraire les données

for game_element in game_elements:
    # Extraire le nom du jeu
    game = game_element.find("a", class_='title').text.strip()
    Name.append(game)
    
    # Extraire la note
    score = game_element.find("div", class_="metascore_w large game positive").text.strip()
    Note.append(score)

    # Extraire la note des joueurs, le .text pose ici problème pour afficher les résultats.
    # On note que le résulat s'affiche entre crochets, il faudra les supprimer par la suite.

    p_score = game_element.find("div", class_="metascore_w user large game positive")
    P_Note.append(p_score)
    
    # Extraire la plateforme

    platform = game_element.find("div", class_="platform").find("span", class_="data").text.strip()
    Platform.append(platform)
    
    # Extraire la date, ici obligation d'utiliser -1 pour séléctionner le dernier span qui nous intéresse, en l'occurence la date

    date = game_element.find("div", class_="clamp-details").find_all('span')[-1].text.strip()
    Date.append(date)

# Créer un DataFrame avec les données extraites
df_metacritic = pd.DataFrame({
    "Name": Name,
    "Platform": Platform,
    "Date": Date,
    "Note MC": Note,
    "Note Players MC": P_Note,
})

#export en csv

#df_metacritic.to_csv("metacritic_data.csv", index=False)

df_metacritic.head(50)

df_metacritic['Note MC'] = df_metacritic['Note MC'].astype(float)

print(df_metacritic[df_metacritic['Note Players MC'].isna()].index)

Name = []
Note = []
P_Note = []
Platform = []
Date = []
page = 0

while page != 80:
    url = f"https://www.metacritic.com/browse/games/score/metascore/all/all/filtered?page={page}"
    session = requests.Session()
    session.headers['User-Agent'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/34.0.1847.131 Safari/537.36'
    response = session.get(url)

    soup = bs(response.content,'html.parser')
    game_elements = soup.findAll("td", class_="clamp-summary-wrap")

    for game_element in game_elements:
        
        game = game_element.find("a", class_='title').text.strip()
        Name.append(game)

        score = game_element.find("div", class_="metascore_w large game positive").text.strip()
        Note.append(score)

        p_score = game_element.find("div", class_="metascore_w user large game positive")
        P_Note.append(p_score)

        platform = game_element.find("div", class_="platform").find("span", class_="data").text.strip()
        Platform.append(platform)

        date = game_element.find("div", class_="clamp-details").find_all('span')[-1].text.strip()
        Date.append(date)

    page += 1
        

# Créer un DataFrame avec les données extraites
df_metacritic = pd.DataFrame({
    "Name": Name,
    "Platform": Platform,
    "Date": Date,
    "Note MC": Note,
    "Note Players MC": P_Note,
})

df_metacritic.head(50)

df_metacritic.info()

df_metacritic['Platform'].value_counts()

df_metacritic['Platform'].nunique()

#Création des listes

Name = []
Note = []
P_Note = []
Platform = []
Date = []
page = 0

#Liste while pour récupérer les 90 premières pages avec la boucle while + for + try en cas d'erreur

while page != 203:
    url = f"https://www.metacritic.com/browse/games/score/metascore/all/all/filtered?page={page}"
    session = requests.Session()
    session.headers['User-Agent'] = 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_2) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/34.0.1847.131 Safari/537.36'
    response = session.get(url)

    soup = bs(response.content,'html.parser')
    game_elements = soup.findAll("td", class_="clamp-summary-wrap")

    for game_element in game_elements:
        try:
            game = game_element.find("a", class_='title')
            if game is not None:
                game = game.text.strip()
            Name.append(game)

            score = game_element.find("div", class_="metascore_w large game positive") or game_element.find("div", class_="metascore_w large game mixed") or game_element.find("div", class_="metascore_w medium game negative")
            if score is not None:
                score = score.text.strip()
            Note.append(score)

            p_score = game_element.find("div", class_="metascore_w user large game positive") or game_element.find("div", class_="metascore_w user large game mixed") or game_element.find("div", class_="metascore_w user medium game negative")
            if p_score is not None:
                p_score = p_score.text.strip()
            P_Note.append(p_score)

            platform = game_element.find("div", class_="platform").find("span", class_="data")
            if platform is not None:
                platform = platform.text.strip()
            Platform.append(platform)

            date = game_element.find("div", class_="clamp-details").find_all('span')[-1]
            if date is not None:
                date = date.text.strip()
            Date.append(date)

        except Exception as e:
            print(f"Erreur détectée sur la page {page}: {e}")

    page += 1
        
# Créer un DataFrame avec les données extraites
df_metacritic = pd.DataFrame({
    "Name": Name,
    "Platform": Platform,
    "Date": Date,
    "Note MC": Note,
    "Note Players MC": P_Note,
})

df_metacritic.head()

Seulement 3 minutes 30 pour scraper l'ensemble des pages les erreurs ralentissaient le bon déroulement du code. Sur 20205 entrées:

- 93% de réussite pour les Notes MC
- 84% de réussite pour les Notes Players MC


df_metacritic.info()

df_metacritic.to_csv("MC_final.csv", index=False)

df = pd.read_csv('MC_final.csv')
df.head(50)


Test Selenium pour scrapper le nom des platform 

import warnings
warnings.filterwarnings("ignore")
from time import sleep
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager
from selenium.webdriver.common.keys import Keys
import pandas as pd
import re
import numpy as np

Créer instance driver pour chrome

try:
    driver = webdriver.Chrome()
except:
    driver = webdriver.Chrome(ChromeDriverManager().install())

accéder à la page

driver.get('https://www.metacritic.com/browse/games/score/metascore/all/all/filtered')

Refuser les cookies

webelement = driver.find_element(by='id', value = "onetrust-reject-all-handler")

driver.execute_script("arguments[0].click();", webelement)

Cliquer sur le premier lien

liens = driver.find_elements_by_class_name("title")

for lien in liens:
    lien.click()

import requests
from bs4 import BeautifulSoup as bs

headers = {
    'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.36'
}

publishers = []
page = 0
while page != 5:
    url = f'https://www.metacritic.com/browse/games/score/metascore/all/all/filtered?page={page}'
    response = requests.get(url, headers=headers)
    soup = bs(response.content, 'html.parser')
    games = soup.find_all('a', class_='title')
    
    for game in games:
        game_url = 'https://www.metacritic.com' + game['href']
        game_response = requests.get(game_url, headers=headers)
        game_soup = bs(game_response.content, 'html.parser')
        publisher = game_soup.find('li', class_='summary_detail publisher').find('span', class_='data').text.strip()
        publishers.append(publisher)
    
    page += 1

print(publishers)

Cela fonctionne mais après réflexion il a été décidé que cette données n'était pas nécéssaire

# Scrapping vgchartz.com

On souhaite scrapper le site suivant: https://www.vgchartz.com/games/games.php?name=&keyword=&console=&region=All&developer=&publisher=&goty_year=&genre=&boxart=Both&banner=Both&ownership=Both&showmultiplat=No&results=200&order=Sales&showtotalsales=0&showpublisher=0&showpublisher=1&showvgchartzscore=0&showvgchartzscore=1&shownasales=0&showdeveloper=0&showcriticscore=0&showcriticscore=1&showpalsales=0&showreleasedate=0&showreleasedate=1&showuserscore=0&showuserscore=1&showjapansales=0&showlastupdate=0&showlastupdate=1&showothersales=0&showshipped=0&showshipped=1

les informations qui nous intéressent sont les suivantes:

Nom du jeu
Plateforme
Editeur
Total Shipped
Release Date

#import des modules de scrapping

from bs4 import BeautifulSoup as bs
from bs4 import UnicodeDammit
import requests
import pandas as pd
import numpy as np
from urllib.request import urlopen



#choisir l'url de VGChartz
url_VG = 'https://www.vgchartz.com/games/games.php?name=&keyword=&console=&region=All&developer=&publisher=&goty_year=&genre=&boxart=Both&banner=Both&ownership=Both&showmultiplat=No&results=200&order=Sales&showtotalsales=0&showpublisher=0&showpublisher=1&showvgchartzscore=0&showvgchartzscore=1&shownasales=0&showdeveloper=0&showcriticscore=0&showcriticscore=1&showpalsales=0&showreleasedate=0&showreleasedate=1&showuserscore=0&showuserscore=1&showjapansales=0&showlastupdate=0&showlastupdate=1&showothersales=0&showshipped=0&showshipped=1'

page_VG = urlopen(url_VG)

soup = bs(page_VG, 'html.parser')


#Création des listes

Name = []
publishers = []
sales = []
release_dates = []
consoles = []


td_elements = soup.find_all('td', style='font-size:12pt;')

# parcourir chaque élément <td> et extraire le texte
for td in td_elements:
    titre = td.text.strip()
    Name.append(titre)

# Extraction des informations pour chaque jeu
for game in soup.find_all('tr', {'style': ['background-image:url(../imgs/chartBar_large.gif); height:70px', 'background-image:url(../imgs/chartBar_alt_large.gif); height:70px', 'font-size:12pt;']}):
    # Extraction du nom du publisher
    publisher = game.find_all('td')[4].text.strip()
    publishers.append(publisher)
    
    # Extraction du nombre de vente
    sale = game.find_all('td')[8].text.strip()
    sales.append(sale)
    
    # Extraction de la date de sortie
    release_date = game.find_all('td')[9].text.strip()
    release_dates.append(release_date)

# Création du dataframe pandas
df = pd.DataFrame({
    'Name': Name,
    'Publisher': publishers,
    'Global Sales': sales,
    'Date de sortie': release_dates
})

# Affichage du dataframe
df

#test pour les Name que nous navons pas réussi à scrapper la première fois
Name = []

td_elements = soup.find_all('td', style='font-size:12pt;')

# parcourir chaque élément <td> et extraire le texte
for td in td_elements:
    titre = td.text.strip()
    Name.append(titre)

df_Name = pd.DataFrame({
    'Name': Name,
})    

df_Name

cela fonctionne, on intègre cela au code final

import pandas as pd
import requests
from bs4 import BeautifulSoup

page = 1
Name = []
publishers = []
sales = []
release_dates = []

while page != 100:
    url = f'https://www.vgchartz.com/games/games.php?page={page}&results=1000&name=&console=&keyword=&publisher=&genre=&order=Sales&ownership=Both&boxart=Both&banner=Both&showdeleted=&region=All&goty_year=&developer=&direction=DESC&month=&year=&sales_type=&headerBackground=&footerBackground='
    response = requests.get(url)
    soup = BeautifulSoup(response.content, 'html.parser')
    
    td_elements = soup.find_all('td', style='font-size:12pt;')
    for td in td_elements:
        titre = td.text.strip()
        Name.append(titre)

    for game in soup.find_all('tr', {'style': ['background-image:url(../imgs/chartBar_large.gif); height:70px', 'background-image:url(../imgs/chartBar_alt_large.gif); height:70px', 'font-size:12pt;']}):
        publisher = game.find_all('td')[4].text.strip()
        publishers.append(publisher)

        sale = game.find_all('td')[8].text.strip()
        sales.append(sale)

        release_date = game.find_all('td')[9].text.strip()
        release_dates.append(release_date)

    page += 1

df = pd.DataFrame({
    'Name': Name,
    'Publisher': publishers,
    'Global Sales': sales,
    'Date de sortie': release_dates
})

df


df.info()

#df.to_csv('VgChartz.csv')

Cleaning du DataFrame

(df['Global Sales'] == 'N/A').value_counts()

#Beaucoup de NA dans la colonne Global Sales, tous les jeux n'ont pas d'informations concernant cette donnée.

df['Publisher'].nunique()

import pandas as pd
import numpy as np
import matplotlib as plt


df = pd.read_csv('VgChartz.csv')
df.head(166)

df.drop('Unnamed: 0', axis=1)

df.drop('Unnamed: 0', axis=1)

df['Global Sales'] = df['Global Sales'].str.replace('m','')

#on retire le "m" de global sales pour n'avoir que des float, on pourra ensuite convertir en float.

df=df.drop('Unnamed: 0', axis=1)

df

df['Global Sales'] = df['Global Sales'].astype(float)
#conversion en float

df.info()

df['Date de sortie'] = pd.to_datetime(df['Date de sortie'])
df['Année de sortie'] = df['Date de sortie'].dt.year
#conversion en format year



df.info()

df = df.drop("Date de sortie",axis=1)


df
