==> 03_data_quality.ipynb <==
import pandas as pd
import numpy as np
import re

# Clean de VG


vg = pd.read_csv('vg_sales_clean.csv')

vg.info()

vg = vg.drop('Unnamed: 0', axis = 1)

# HOMOGENEISATION DE LA CASSE

cols = ['Name','Platform','Genre','Publisher']
for col in cols :
  vg[col] = vg[col].apply(lambda x : x.upper())

# RETRAIT DES ELEMENTS DE TYPE (a-zA-Z0-9) DANS LES TITRES DE JEUX

for i, element in enumerate(vg['Name']):
  if '(' in element :
    r1 = re.findall(r"\(.*?\)", element)
    vg['Name'][i] = element.replace(r1[0], '').strip()

vg.head()

#vg.to_csv('vg.csv', index = False)

# Clean de MC

mc = pd.read_csv('MC_final.csv')
mc = mc.drop('Date', axis = 1)

mc.info()

# MODIFICATION NOMS DES COLONNES

mc = mc.rename(columns = {'Note MC':'Test_MC','Note Players MC':'Players_MC'})

# REGROUPEMENT PAR PLATEFORMES

platform_replace = {'PlayStation 4' : 'Playstation', 'Xbox 360' : 'Xbox','PlayStation 2' : 'Playstation',
                    'PlayStation 3' : 'Playstation', 'PlayStation 5' : 'Playstation','Xbox One' : 'Xbox',
                    'PSP':'PlayStation','Game Boy Advance':'GB','PlayStation Vita' : 'Playstation','3DS':'DS',
                    'GameCube':'Autre','Xbox Series X':'Xbox','Wii U':'Wii', 'Dreamcast':'Autre',
                    'Nintendo 64':'Nintendo','Stadia':'Autre'}

mc['Platform'] = mc['Platform'].replace(platform_replace)

# HOMOGENEISATION DE LA CASSE

mc['Name'] = mc['Name'].apply(lambda x: x.upper())
mc['Platform'] = mc['Platform'].apply(lambda x: x.upper())

# RETRAIT DES ELEMENTS DE TYPE (a-zA-Z0-9) DANS LES TITRES DE JEUX

for i, element in enumerate(mc['Name']):
  if '(' in element :
    r1 = re.findall(r"\(.*?\)", element)
    mc['Name'][i] = element.replace(r1[0], '').strip()

# GROUPBY NAME et PLATFORM : PS1 = Playstation ; PS4 = Playstation pour un même jeu

mc = mc.groupby(['Name','Platform']).mean().reset_index()
mc.head()

print(mc.info())
print('\n')
print(mc.isna().sum())

#mc.to_csv('/content/drive/MyDrive/Projet_Gaming_Sales/CODE_FINAL/csv/mc.csv', index = False)

# Clean de JV

jv = pd.read_csv('dataframe_jeuxvideos.csv')
jv = jv.drop('Unnamed: 0', axis = 1)
display(jv.head())
print(jv.shape)

# MODIFICATION NOMS DES COLONNES
jv = jv.rename(columns = {'titres' : 'Name', 
                          'test':'Test_JV', 
                          'avis':'Players_JV'})

# HOMOGENEISATION DE LA CASSE
jv['Name'] = jv['Name'].apply(lambda x : x.upper())

# RETRAIT DES ELEMENTS DE TYPE (a-zA-Z0-9) DANS LES TITRES DE JEUX
for i, element in enumerate(jv['Name']):
  if '(' in element :
    r1 = re.findall(r"\(.*?\)", element)
    jv['Name'][i] = element.replace(r1[0], '').strip()

# MODIFICATION DES VALEURS DES NOTES

to_replace = {'- -/20' : '0/20', 'test non renseigné' : '0/20',             # remplacement des notes non renseignées
              'avis non renseigné' : '0/20','Non renseigné':'0/20',         # par des 0 (elles deviendronts des NaN à terme)
              '18,5/20':'18.5/20'}                                           

jv['Test_JV'] = jv['Test_JV'].replace(to_replace)
jv['Players_JV'] = jv['Players_JV'].replace(to_replace)

jv['Test_JV'] = jv['Test_JV'].apply(lambda x : x[:-3]).astype(float)        # on tronque la partie '/20' des notes : 18/20 => 18
jv['Players_JV'] = jv['Players_JV'].apply(lambda x : x[:-3]).astype(float)

to_replace_nan = {0 : np.nan}                                               # On remplace les 0 par des NaN pour rester fidèle à la réalité
jv['Test_JV'] = jv['Test_JV'].replace(to_replace_nan)
jv['Players_JV'] = jv['Players_JV'].replace(to_replace_nan)

# HOMOGENEISATION DES TITRES : POSITION DU CARACTERE ':'

for i, element in enumerate(jv['Name']) :
  if ' : ' in element :
    r2 = re.findall(r"\ :.*?\ ", element)
    jv['Name'][i] = element.replace(r2[0], ': ')

# RECUPERATION DES PLATEFORMES DANS LES TITRES DES JEUX

plateforme_jv = []
for element in jv['Name']:
  if ' SUR ' in element :
    plateforme_jv.append(element.split(' SUR ')[1])
  else :
    plateforme_jv.append(np.nan)

jv['Platform'] = plateforme_jv

plateforme_replace = {'PS1':'PLAYSTATION','PS2':'PLAYSTATION','PS3':'PLAYSTATION','PS4':'PLAYSTATION',
                      'PS5':'PLAYSTATION','PSP':'PLAYSTATION','VITA':'PLAYSTATION','XBOX SERIES':'XBOX',
                      '3DS':'DS','360':'XBOX','WIIU':'WII','GBA':'GB','IOS':'MOBILE','ONE':'XBOX','NGC':'GB',
                      'N64':'NINTENDO','NES':'NINTENDO','SNES':'NINTENDO','WEB':'MOBILE','GBC':'GB','INTERNET':'MOBILE',
                      'DCAST':'AUTRE','MD':'AUTRE','SATURN':'AUTRE','NGAGE':'AUTRE','MS':'AUTRE','LINUX':'MOBILE','NEW 3DS':'DS'}


# MODIFICATION DES TITRES

jv['Name'] = jv['Name'].apply(lambda x : x.split(' SUR ')[0])

jv.head()

for cols in ['Name','Test_JV','Players_JV','Platform']:
  print('Nombre de NaN dans',cols,' : ',jv[cols].isna().sum(), 
        '\n', '% de NaN dans',cols, ' : ',(100 * jv[cols].isna().sum() / len(jv[cols])).round(2),'%')

jv = jv.drop('Platform', axis = 1)

#jv.to_csv('jv.csv', index = False)

# Clean de VGChartz

## Limitation de la colonne Platform

vgc = pd.read_csv('vg_chartz_cleaned.csv')
vgc.head()

vgc.info()

vgc = vgc.rename(columns = {'Global Sales':'Global_Sales','Date de sortie':'Year'})

for i, element in enumerate(vgc['Name']) :
  if '(' in element :
    r2 = re.findall(r"\(.*?\)", element)
    vgc['Name'][i] = element.replace(r2[0], '').strip()

for i, element in enumerate(vgc['Name']) :
  if 'Read the review' in element :
    r3 = re.findall("Read the review", element)
    vgc['Name'][i] = element.replace(r3[0], '').strip()

vgc = vgc.groupby(['Name','Platform','Publisher']).agg({'Global_Sales':'sum','Year':'mean'}).sort_values(by = 'Global_Sales', ascending = False).reset_index()

vgc['Global_Sales'] = vgc['Global_Sales'].replace({0.0:np.nan})
vgc = vgc.dropna()

for col in ['Name','Platform','Publisher']:
  vgc[col] = vgc[col].apply(lambda x : x.upper())

vgc.head()

vgc.info()

# Illustration de la limite de la colonne platform

vgc[vgc['Name'].str.contains('GRAND THEFT AUTO')]

# Autre illustration de la non pertinence de ces données sur un 2eme DataFrame issu d'une autre tentative de scrapping

vgc2 = pd.read_csv('dataframe_vgc.csv')
vgc2.head()

vgc2['Total_sales'] = vgc2['Total_sales'].fillna('0.00m')
vgc2['Total_shipped'] = vgc2['Total_shipped'].fillna('0.00m')

vgc2['Total_sales'] = vgc2['Total_sales'].apply(lambda x : x[:-1]).astype(float)
vgc2['Total_shipped'] = vgc2['Total_shipped'].apply(lambda x : x[:-1]).astype(float)

nan_replace = {0.00:np.nan}
vgc2['Total_sales'] = vgc2['Total_sales'].replace(nan_replace)
vgc2['Total_shipped'] = vgc2['Total_shipped'].replace(nan_replace)

vgc2['Date'] = pd.to_datetime(vgc2['Date']).dt.year

for i, element in enumerate(vgc2['Name']) :
  if '(' in element :
    r2 = re.findall(r"\(.*?\)", element)
    vgc2['Name'][i] = element.replace(r2[0], '').strip()

for i, element in enumerate(vgc2['Name']) :
  if 'Read the review' in element :
    r3 = re.findall("Read the review", element)
    vgc2['Name'][i] = element.replace(r3[0], '').strip()

for i in ['Name','Publisher']:
  vgc2[i] = vgc2[i].apply(lambda x : x.upper())


vgc2 = vgc2.drop(["Total_shipped","Date"], axis = 1)

# Total_shipped contient tres peu de données et Date en contient beaucoup mais en vue de faire un dropna(), on supprime Date
# pour éviter de perdre trop d'informations sur le futur dropna()

vgc2 = vgc2.dropna()
vgc2.head()

vgc2 = vgc2.groupby(['Name','Publisher']).sum().sort_values(by = 'Total_sales', ascending = False).reset_index()
vgc2.head()

# données de ventes mais pas de consoles, on ne sait pas la proportion des ventes de chaque consoles
# informations peu voir pas exploitables

## Autre illustration de la limitation de ce jeu de données

# récupération des plateformes à la suite d'un nouveau scrapping du site pour récupérer les consoles uniquement

console = pd.read_csv('Consoles-vgchartz.csv')
console = console.drop(['Unnamed: 0.1', 'Unnamed: 0'], axis = 1)
console = console.rename(columns = {'consoles':'Platform', 'games;':'Name'})
console = console.dropna()
console['Name'] = console['Name'].apply(lambda x : x[:-1])
platform_df = console.iloc[:20000] # uniquement des NaN au dela
display(console.head())
vgc_df = pd.read_csv('dataframe_vgc.csv')


df1 = vgc_df.merge(platform_df, on = 'Name', how = 'left')
df1.head()

df1['Total_sales'] = df1['Total_sales'].fillna('0.00m')
df1['Total_shipped'] = df1['Total_shipped'].fillna('0.00m')

df1['Total_sales'] = df1['Total_sales'].apply(lambda x : x[:-1]).astype(float)
df1['Total_shipped'] = df1['Total_shipped'].apply(lambda x : x[:-1]).astype(float)

nan_replace = {0.00:np.nan}
df1['Total_sales'] = df1['Total_sales'].replace(nan_replace)
df1['Total_shipped'] = df1['Total_shipped'].replace(nan_replace)

df1['Date'] = pd.to_datetime(df1['Date']).dt.year

for i, element in enumerate(df1['Name']) :
  if '(' in element :
    r2 = re.findall(r"\(.*?\)", element)
    df1['Name'][i] = element.replace(r2[0], '').strip()

for i, element in enumerate(df1['Name']) :
  if 'Read the review' in element :
    r3 = re.findall("Read the review", element)
    df1['Name'][i] = element.replace(r3[0], '').strip()

df1.head()

df5 = df1 # travail sur une copie
df5 = df5.drop('Total_shipped', axis = 1)
df5 = df5.dropna()
df5.info()

# Illustration de la mauvaise information induite par la platform
df5[df5['Name'].str.contains('Call of Duty')]

# Fusion et gestion des valeurs manquantes

# Importation des DataFrame clean issus de la Data Quality

vg = pd.read_csv("vg.csv")
mc = pd.read_csv("mc.csv")
jv = pd.read_csv("jv.csv")

# Merge des dataframes selon différentes combinaison

mc_jv = jv.merge(mc, on = 'Name', how = 'left')
vg_mc = vg.merge(mc, on = ['Name','Platform'], how = 'left')
vg_jv = vg.merge(jv, on = 'Name', how = 'left')
vg_mc_jv = vg.merge(mc_jv, on = ['Name','Platform'], how = 'left')

# Exportation csv

#mc_jv.to_csv('/content/drive/MyDrive/Projet_Gaming_Sales/CODE_FINAL/csv/metacritic_jeuxvideos.csv', index = False)
#vg_mc.to_csv('/content/drive/MyDrive/Projet_Gaming_Sales/CODE_FINAL/csv/vg_metacritic.csv', index = False)
#vg_jv.to_csv('/content/drive/MyDrive/Projet_Gaming_Sales/CODE_FINAL/csv/vg_jeuxvideos.csv', index = False)
#vg_mc_jv.to_csv('/content/drive/MyDrive/Projet_Gaming_Sales/CODE_FINAL/csv/vg_metaritic_jeuxvideos.csv', index = False)

# Etude des NAN - DataFrame mc_jv

display(mc_jv.head())
for i in mc_jv.columns :
  print(i,':',mc_jv[i].isna().sum(),'NaN =', (100*mc_jv[i].isna().sum()/len(mc_jv[i])).round(2),'%')
  
#instanciation d'une boucle "for" pour afficher le pourcentage de NaN au sein du dataframe. 
#Un pourcentage plutôt élevé, en effet bien que les sites regroupent la quasi totalité des jeux existants, rares sont ceux qui possèdent une note "presse" et "joueurs".
#Cela ne représente pas un problème en soi puisque nous cherchons ici à rajouter des jeux plus récents à notre DF d'origine.

# Etude des NAN - DataFrame vg_jv

display(vg_jv.head())
for i in vg_jv.columns :
  print(i,':',vg_jv[i].isna().sum(),'NaN : ', (100*vg_jv[i].isna().sum()/len(vg_jv[i])).round(2),'%')
  
#Ici, on rajoute les notes du site jeuxvideo.com à notre dataframe d'origine, le nombre de NaN est légérement réduit.

# Etude des NAN - DataFrame vg_mc

display(vg_mc.head())
for i in vg_mc.columns :
  print(i,':',vg_mc[i].isna().sum(),'NaN = ', (100*vg_mc[i].isna().sum()/len(vg_mc[i])).round(2),'%')
  
#Même chose avec le site MetaCritic, premier constat, beaucoup moins de valeurs NaN qu'avec jeuxvideo.com

# Etude des NAN - DataFrame mc_jv

display(vg_mc_jv.head())
for i in vg_mc_jv.columns :
  print(i,':',vg_mc_jv[i].isna().sum(),'NaN = ', (100*vg_mc_jv[i].isna().sum()/len(vg_mc_jv[i])).round(2),'%')

vg_mc_jv = vg_mc_jv.dropna()

#vg_mc_jv.to_csv("/content/drive/MyDrive/Projet_Gaming_Sales/CODE_FINAL/csv/vg_mc_jv.csv", index = False)
# DF que l'on décide de garder pour la modélisation
